Updated collections

+++ bin/ku4js-data-uncompressed.js
@@ -1318,6 +1318,7 @@ function imageFileField(selector) {
 
 imageFileField.prototype = {
     maxDims: function(value) { return this.property("maxDims", $.point.parse(value)); },
+    $write: function(){ return; },
     $readFiles: function(func, scp) {
         var files = $.list(this.files()),
             maxDims = this._maxDims,
@@ -1602,7 +1603,7 @@ collection.prototype = {
     isEmpty: function() { return this._data.isEmpty(); },
     count: function() { return this._data.count(); },
     store: function(store) { this._store = collection_getStore(store); return this; },
-    save: function(callback) { this._store.write(this, callback); return this; },
+    save: function(callback, scope) { this._store.write(this, callback, scope); return this; },
     init: function(list) { return this.remove().insertList(list); },
     find: function(query) {
         if(!$.exists(query)) return this._data.values();
@@ -1841,24 +1842,24 @@ function abstractStore(name) {
     this._name = name || "ku4indexedDbStore";
 }
 abstractStore.prototype = {
-    read: function(collectionName, callback) {
+    read: function(collectionName, callback, scope) {
         callback(new Error("Not Implemented"), null);
         return this;
     },
-    write: function(collection, callback) {
+    write: function(collection, callback, scope) {
         callback(new Error("Not Implemented"), null);
         return this;
     },
-    remove: function(collection, callback) {
+    remove: function(collection, callback, scope) {
         callback(new Error("Not Implemented"), null);
         return this;
     },
-    __delete: function(callback) {
+    __delete: function(callback, scope) {
         callback(new Error("Not Implemented"), null);
         return this;
     },
-    __reset: function(callback) {
-        return this.__delete(callback);
+    __reset: function(callback, scope) {
+        return this.__delete(callback, scope);
     }
 };
 
@@ -1866,8 +1867,10 @@ function indexedDbStore(name) {
     this._name = name || "ku4indexedDbStore";
 }
 indexedDbStore.prototype = {
-    read: function(collectionName, callback) {
+    read: function(collectionName, callback, scope) {
+        if(!$.isFunction(callback)) throw new $.ku4exception("$.ku4indexedDbStore", "Invalid callback parameter at read");
         var name = this._name,
+            scp = scope || this,
             me = this;
 
         ku4indexedDbStore_openDb(name, function (err, db) {
@@ -1877,69 +1880,83 @@ indexedDbStore.prototype = {
                 .onsuccess = function (event) {
                     var data = event.target.result,
                         collection = $.ku4collection(collectionName, data).store(me);
-                    if($.exists(callback)) callback(null, collection);
+                    if($.exists(callback)) callback.call(scp, null, collection);
                     db.close();
                 };
         }, collectionName);
         return this;
     },
-    write: function(collection, callback) {
+    write: function(collection, callback, scope) {
+        if(!$.isFunction(callback)) throw new $.ku4exception("$.ku4indexedDbStore", "Invalid callback parameter at write");
         var collectionName = collection.name(),
             name = this._name,
+            scp = scope || this,
             me = this;
 
         ku4indexedDbStore_openDb(name, function (err, db) {
             if($.exists(err)) {
-                if($.exists(callback)) callback(err, null);
+                if($.exists(callback)) callback.call(scp, err, me, collection);
             }
             else {
                 var request = db.transaction([collectionName], "readwrite").objectStore(collectionName).put(collection.toObject(), 1);
                 request.onerror = function () {
-                    if($.exists(callback)) callback(new Error("Error writing data to indexedDbStore"), me);
+                    if($.exists(callback)) callback.call(scp, new Error("Error writing data to indexedDbStore"), me, collection);
                     db.close();
                 };
                 request.onsuccess = function () {
-                    if($.exists(callback)) callback(null, me);
+                    if($.exists(callback)) callback.call(scp, null, me, collection);
                     db.close();
                 };
             }
         }, collectionName);
         return this;
     },
-    remove: function(collection, callback) {
+    remove: function(collection, callback, scope) {
+        if(!$.isFunction(callback)) throw new $.ku4exception("$.ku4indexedDbStore", "Invalid callback parameter at remove");
         var collectionName = collection.name(),
             name = this._name,
+            scp = scope || this,
             me = this;
 
         ku4indexedDbStore_openDb(name, function (err, db) {
-            if($.exists(err)) callback(err, null);
+            if($.exists(err)) callback.call(scp, err, me, collection);
             else {
                 var request = db.transaction([collectionName], "readwrite").objectStore(collectionName)["delete"](1);
                 request.onerror = function () {
-                    if($.exists(callback)) callback(new Error("Error removing data to indexedDbStore"), me);
+                    if($.exists(callback)) callback.call(scp, new Error("Error removing data to indexedDbStore"), me, collection);
                     db.close();
                 };
                 request.onsuccess = function () {
-                    if($.exists(callback)) callback(null, me);
+                    if($.exists(callback)) callback.call(scp, null, me, collection);
                     db.close();
                 };
             }
         }, collectionName);
         return this;
     },
-    __delete: function(callback) {
+    __delete: function(callback, scope) {
+        if(!$.isFunction(callback)) throw new $.ku4exception("$.ku4indexedDbStore", "Invalid callback parameter at __delete");
         var idxdb = indexedDB || webkitIndexedDB || mozIndexedDB,
             request = idxdb.deleteDatabase(this._name),
+            scp = scope || this,
             me = this;
 
-        request.onerror = function() { if($.exists(callback)) callback(new Error("Error deleting indexedDbStore.", me))};
-        request.onsuccess = function() { if($.exists(callback)) callback(null, me); };
+        request.onerror = function() {
+            if($.exists(callback))
+                callback.call(scp, new Error("Error deleting indexedDbStore.", me))}
+        ;
+        request.onsuccess = function() {
+            if($.exists(callback))
+                callback.call(scp, null, me);
+        };
         return this;
     },
-    __reset: function(callback) {
+    __reset: function(callback, scope) {
+        if(!$.isFunction(callback)) throw new $.ku4exception("$.ku4indexedDbStore", "Invalid callback parameter at __reset");
+        var scp = scope || this;
         this.__delete(function(err, store) {
             __ku4indexedDbStoreVersion = 0;
-            if($.exists(callback)) callback(err, store);
+            if($.exists(callback)) callback.call(scp, err, store);
         });
     }
 };
@@ -2005,28 +2022,32 @@ function ku4indexedDbStore_getIdbx()
 
 function localStorageStore() { }
 localStorageStore.prototype = {
-    read: function(collectionName, callback) {
+    read: function(collectionName, callback, scope) {
         var collection = localStorage.getItem(collectionName),
             ku4collection =  ($.exists(collection))
                 ? $.ku4collection.deserialize(collection).store(this)
-                : $.ku4collection(collectionName).store(this);
+                : $.ku4collection(collectionName).store(this),
+            scp = scope || this;
 
-        if($.exists(callback)) callback(null, ku4collection);
+        if($.exists(callback)) callback.call(scp, null, ku4collection);
         return ku4collection;
     },
-    write: function(collection, callback) {
+    write: function(collection, callback, scope) {
+        var scp = scope || this;
         localStorage.setItem(collection.name(), collection.serialize());
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this, collection);
         return this;
     },
-    remove: function(collection, callback) {
+    remove: function(collection, callback, scope) {
+        var scp = scope || this;
         localStorage.removeItem(collection.name());
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this, collection);
         return this;
     },
-    __delete: function(callback) {
+    __delete: function(callback, scope) {
+        var scp = scope || this;
         localStorage.clear();
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this);
         return this;
     }
 };
@@ -2037,28 +2058,32 @@ $.ku4localStorageStore = function() { return new localStorageStore(); };
 var __ku4MemoryStore = $.dto();
 function memoryStore() { }
 memoryStore.prototype = {
-    read: function(collectionName, callback) {
+    read: function(collectionName, callback, scope) {
         var collection = __ku4MemoryStore.find(collectionName),
             ku4collection =  ($.exists(collection))
                 ? $.ku4collection.deserialize(collection).store(this)
-                : $.ku4collection(collectionName).store(this);
+                : $.ku4collection(collectionName).store(this),
+            scp = scope || this;
 
-        if($.exists(callback)) callback(null, ku4collection);
+        if($.exists(callback)) callback.call(scp, null, ku4collection);
         return ku4collection;
     },
-    write: function(collection, callback) {
+    write: function(collection, callback, scope) {
+        var scp = scope || this;
         __ku4MemoryStore.update(collection.name(), collection.serialize());
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this, collection);
         return this;
     },
-    remove: function(collection, callback) {
+    remove: function(collection, callback, scope) {
+        var scp = scope || this;
         __ku4MemoryStore.remove(collection.name());
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this, collection);
         return this;
     },
-    __delete: function(callback) {
+    __delete: function(callback, scope) {
+        var scp = scope || this;
         __ku4MemoryStore.clear();
-        if($.exists(callback)) callback(null, this);
+        if($.exists(callback)) callback.call(scp, null, this);
         return this;
     }
 };
